# ENH: speedup of abeles function