--- a/src/reflect.jl
+++ b/src/reflect.jl
@@ -1,13 +1,16 @@
-using Interpolations, Distributions, ForwardDiff, LinearAlgebra
+using Interpolations, Distributions, ForwardDiff, LinearAlgebra, FastGaussQuadrature
 
 const TINY = 1e-30
 const _FWHM = 2 * sqrt(2 * log(2.0))
 const PI4 = 4e-6 * pi
+const _INTLIMIT = 3.5
 
 
-Base.float(d::ForwardDiff.Dual{T}) where T = ForwardDiff.Dual{T}(float(d.value), d.partials)
-Base.prevfloat(d::ForwardDiff.Dual{T,V,N}) where {T,V,N} = ForwardDiff.Dual{T}(prevfloat(float(d.value)), d.partials)
-Base.nextfloat(d::ForwardDiff.Dual{T,V,N}) where {T,V,N} = ForwardDiff.Dual{T}(nextfloat(float(d.value)), d.partials)
+# Base.float(d::ForwardDiff.Dual{T}) where T = ForwardDiff.Dual{T}(float(d.value), d.partials)
+# Base.prevfloat(d::ForwardDiff.Dual{T,V,N}) where {T,V,N} = ForwardDiff.Dual{T}(prevfloat(float(d.value)), d.partials)
+# Base.nextfloat(d::ForwardDiff.Dual{T,V,N}) where {T,V,N} = ForwardDiff.Dual{T}(nextfloat(float(d.value)), d.partials)
+
+
 function Base.ldexp(x::T, e::Integer) where T<:ForwardDiff.Dual
     if e >=0
         x * (1<<e)
@@ -15,6 +18,7 @@ function Base.ldexp(x::T, e::Integer) where T<:ForwardDiff.Dual
         x / (1<<-e)
     end
 end
+
 
 """
     abeles(q::Array{Float64, 1}, layers::Array{Float64, 2})
@@ -32,57 +36,56 @@ function abeles(q, w)
     nlayers = size(w, 1) - 2
     npnts = length(q)
 
-    # SLD in each layer, relative to first.
-    sld = zeros(ComplexF64, nlayers + 2)
-    thickness = zeros(ComplexF64, nlayers)
-    rough_sqr = Vector{ComplexF64}(undef, (nlayers + 1))
-    reflectivity = Vector{ComplexF64}(undef, (npnts))
+    reflectivity = Vector{Any}(undef, (npnts))
     oneC = Complex(1.0)
 
-    for i = 2:nlayers+2
-        sld[i] = ((w[i, 2] - w[1, 2]) + ((abs(w[i, 3]) + TINY))im) * pi * 4.0e-6
-        rough_sqr[i - 1] = -2.0 * w[i, 4]^2
-    end
-    for i=1:nlayers
-        thickness[i] = Complex(0, abs(w[i+1, 1]))
-    end
-    MRtotal = Array{ComplexF64}(undef, (2, 2))
-    MI = Array{ComplexF64}(undef, (2, 2))
-    temp = similar(MRtotal)
-
     for j in eachindex(q)
-        qq2 = (q[j] * q[j] / 4.0)
-        kn = (q[j] / 2.)
+        qq2 = (q[j] * q[j] / 4.0) + 0.0im
+        kn = (q[j] / 2.) + 0.0im
+
+        # variables are local to if blocks
+        local MRtotal11, MRtotal12, MRtotal21, MRtotal22
 
         for i = 1:nlayers+1
             # wavevector in the layer
-            kn_next = sqrt(qq2 - sld[i + 1])
+            sld_next = ((w[i+1, 2] - w[1, 2]) + ((abs(w[i+1, 3]) + TINY))im) * pi * 4.0e-6
+            kn_next = sqrt(qq2 - sld_next)
+
             # reflectance of the interface
-            rj = (kn - kn_next)/(kn + kn_next) * exp(kn * kn_next * rough_sqr[i])
+            rj = (kn - kn_next)/(kn + kn_next) * exp(kn * kn_next * (-2.0 * w[i+1, 4]^2))
 
             if i == 1
                 # characteristic matrix for first interface
-                MRtotal[1, 1] = oneC
-                MRtotal[1, 2] = rj
-                MRtotal[2, 2] = oneC
-                MRtotal[2, 1] = rj
+                MRtotal11 = oneC
+                MRtotal12 = rj
+                MRtotal21 = rj
+                MRtotal22 = oneC
             else
                 # work out the beta for the layer
-                beta = exp(kn * thickness[i - 1])
+                beta = exp(kn * (abs(w[i, 1]) * 1im))
+
                 # this is the characteristic matrix of a layer
-                MI[1, 1] = beta
-                MI[1, 2] = rj * beta
-                MI[2, 2] = oneC / beta
-                MI[2, 1] = rj * MI[2, 2]
-                # multiply MRtotal, MI to get the updated total matrix.
-                # do an in-place multiplication to reduce allocations
-                mul!(temp, MRtotal, MI)
-                MRtotal .= temp
-                # MRtotal = MRtotal * MI
+                MI11 = beta
+                MI12 = rj * beta
+                MI22 = oneC / beta
+                MI21 = rj * MI22
+
+                # propagate optical matrix by matmul
+                p11 = MRtotal11 * MI11 + MRtotal12 * MI21
+                p12 = MRtotal11 * MI12 + MRtotal12 * MI22
+                p21	= MRtotal21 * MI11 + MRtotal22 * MI21
+                p22 = MRtotal21 * MI12 + MRtotal22 * MI22
+
+                MRtotal11 = p11
+                MRtotal12 = p12
+                MRtotal21 = p21
+                MRtotal22 = p22
+
             end
             kn = kn_next;
+
         end
-        reflectivity[j] = MRtotal[2, 1] / MRtotal[1, 1]
+        reflectivity[j] = MRtotal21 / MRtotal11
     end
     return real(reflectivity .* conj(reflectivity))
 end
@@ -156,7 +159,7 @@ end
 
 Perform the reflectometry calculation with a pointwise smearing.
 
-### Parmaeters
+### Parameters
 - `q::Array{Float64, 1}` : q-vector values.
 - `w::Array{Any, 2}` : an Nx4 array, where N is the number of layers in the system, 1st item in a given row is the thickness, the 2nd the SLD, the 3rd the imaginary SLD, and the 4th the roughness with the layer above.
 - `dq::Array{Float64, 1}` : `dq` values for each `q` value.
@@ -165,8 +168,41 @@ Perform the reflectometry calculation with a pointwise smearing.
 ### Returns
 - `::Array{Float64, 1}` : smeared reflectometry values for the given q-vectors, with a pointise smearing.
 """
-function pointwise_smearing(q::Array{Float64, 1}, w::Array{Any, 2}, dq::Array{Float64, 1}; quad_order::Int32=17)
+function pointwise_smearing(q, w, dq, quad_order::Int=17)
+    # not sure how one checks that q, dq have same size.
+    npnts = length(q)
+
+    # get the gauss-legendre weights and abscissae
+    # TODO: is it possible to use a LRU cache?
+    abscissa, weights = gausslegendre(quad_order)
+
+    # get the normal distribution at that point
+    prefactor = 1.0 / sqrt(2 * pi)
+
+    function gauss(x)
+        return exp(-0.5 * x * x)
+    end
+
+    # TODO: is it possible to use a LRU cache?
+    gaussvals = prefactor * gauss.(abscissa * _INTLIMIT) .* weights
+
+    # integration between -3.5 and 3.5 sigma
+    va = q .- _INTLIMIT .* dq ./ _FWHM
+    vb = q .+ _INTLIMIT .* dq ./ _FWHM
+
+    # (quad_order, npnts)
+    qvals_for_res = [(ab * (li[2] - li[1]) + li[2] + li[1])/2.0 for ab=abscissa, li=zip(va, vb)]
+    smeared_rvals = abeles(qvals_for_res, w)
+
+    # abeles flattens the q vector
+    smeared_rvals = reshape(smeared_rvals, size(qvals_for_res))
+    m = reshape(gaussvals, (quad_order, 1))
+
+    smeared_rvals .*= m
+    rvals = sum(smeared_rvals, dims=1) .* _INTLIMIT
+    return reshape(rvals, size(q))
 end
+
 
 """
     constant_smearing(q::Array{Float64, 1}, w::Array{Any, 2}, resolution::Any, scale::Any, bkg::Any)