--- a/test/data/test4.txt
+++ b/test/data/test4.txt
@@ -1,0 +1,6 @@
+# a four layer system that has an absorption
+# has been resolution smeared to dQ/Q = 0.05 FWHM or 0.0212 1-sigma
+# generated using refnx with pointwise + 10001 points in the gaussian
+# quadrature
+layers/test0.layers
+data/test4.dat