I'm also interested in learning a bit more about Julia, and I just found your code. I rewrote the refnx refcalc.cpp in [different ways](https://gist.github.com/andyfaff/cc26b87da9eaf47833c3d99c4f6b7e68), and came up with this PR.
I got the following outputs from benchmarking the existing code, and my proposed modifications. The two are surprisingly similar. There is no threading in here, I didn't find it gave a speedup. Hopefully the code should still work with `ForwardDiff`.

```
>>> using BenchmarkTools
>>> w = [[0 2.07 0 0]
     [100 3.47 0 3]
     [500 -0.5 0.00001 3]
     [0 6.36 0 3]]

>>> q = Array(LinRange(0.01, 0.5, 1000))
>>> # existing Jeff code
>>> @benchmark abeles(q, w)
BenchmarkTools.Trial: 
  memory estimate:  2.22 MiB
  allocs estimate:  68016
  --------------
  minimum time:     1.571 ms (0.00% GC)
  median time:      1.742 ms (0.00% GC)
  mean time:        1.996 ms (11.30% GC)
  maximum time:     47.865 ms (95.98% GC)
  --------------
  samples:          2502
  evals/sample:     1

>>> # using this PR
>>> @benchmark abeles3(q, w)
BenchmarkTools.Trial: 
  memory estimate:  55.98 KiB
  allocs estimate:  10
  --------------
  minimum time:     323.171 μs (0.00% GC)
  median time:      352.502 μs (0.00% GC)
  mean time:        376.364 μs (2.54% GC)
  maximum time:     51.587 ms (99.27% GC)
  --------------
  samples:          10000
  evals/sample:     1
```