@andyfaff 2020-12-14
I did a quick check of pointwise smearing against refnx and the two match up. I didn't add a test here though.
To calculate the Gauss Legendre weights another package, `FastGaussQuadrature` is needed. I'm not sure how to add the package to Project.toml, that's why the tests are failing.