@andyfaff 2020-12-14
I added `FastGaussQuadrature` to `Project.toml` by:

- `activate .`
- `update`
- `add FastGaussQuadrature`

in `pkg` mode in the Julia REPL. This seems to have updated a lot of stuff in the manifest. Please let me know if you'd like me to back that out.