@andyfaff 2020-12-14
Whilst this abeles seems to be a lot faster I don't think it's going to work with ForwardDiff immediately, because I'm using `Vector{ComplexF64}(sz)` instead of `Vector{Any}(sz)`. Apparently some of these issues are fixed in nightly builds of Julia/most recent ForwardDiff.
I'll see if I can get things to work with that.