@andyfaff 2020-12-15
Ok, it turns out that it's a bad idea to have any `Vector{Any}`. Whilst they're needed for `ForwardDiff` they slow down the reflectivity calculation as there are lots of allocations.

aec7901 removes all `Vector{Any}` instances replacing with straight variables. Local testing shows the reflectivity calculation is still fast, and that ForwardDiff doesn't complain. Not complaining doesn't mean that I've tested the gradient against a finite difference differentiation, just that calculating an autodiffed gradient works.