@arm61 2020-12-15
What is the `@benchmark` without the `Vector{Any}`?