@andyfaff 2020-12-15
```
>>> w = [[0 2.07 0 0]
     [100 3.47 0 3]
     [500 -0.5 0.00001 3]
     [0 6.36 0 3]]

>>> q = Array(LinRange(0.01, 0.5, 1000))
>>>@benchmark abeles(q, w)

BenchmarkTools.Trial: 
  memory estimate:  187.94 KiB
  allocs estimate:  4993
  --------------
  minimum time:     435.128 μs (0.00% GC)
  median time:      439.698 μs (0.00% GC)
  mean time:        461.373 μs (3.52% GC)
  maximum time:     39.935 ms (98.87% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

Not as good as I first had, but should differentiate ok. Before  it's merged you should look over the toml files. I don't think I quite have the full hang of that aspect yet.