@andyfaff 2020-12-16
Added ORSO tests for abeles kernel, including absorption effects. 

Test `pointwise_smearing` calculation. The tolerances on this are quite loose at the moment (rtol=0.03), it may warrant more investigation of the pointwise smearing calculation.