@andyfaff 2020-12-16
Add perform a better `constant_smearing` check