@arm61 2020-12-17
Very nice PR. And a great speed up too!